requirements.txt
langchain-openai
langchain-core
langchain-community
langchain-ollama
streamlit
python-dotenv